+++ frontend/app/api/admin/diagnostics/suspended/route.ts
@@ -2,7 +2,7 @@ import { Role } from "@prisma/client";
 import { NextResponse } from "next/server";
 import { requireRole } from "../../../../../services/auth";
 import { getUnresolvedSuspension } from "../../../../../util/suspension";
-import { formatETDateString } from "../../../../../util/timeUtils";
+import { formatETDateString } from "../../../../../util/date/timeUtils";
 import { prisma } from "../../../../../lib/prisma";
 
 const notDeleted = { deletedAt: null };
