+++ frontend/app/api/update/meeting/resume/route.ts
@@ -2,7 +2,7 @@ import { Prisma, Role, SuspensionPeriod } from '@prisma/client';
 import { NextResponse, after } from 'next/server';
 import { requireRole } from '../../../../../services/auth';
 import { deleteCalendarEvent, createCalendarEvent, calendarIdsForMeeting } from '../../../../../services/googleCalendar';
-import { formatETDateString } from '../../../../../util/timeUtils';
+import { formatETDateString } from '../../../../../util/date/timeUtils';
 import {
   getUnresolvedSuspension,
   reconcilePendingResume,
