+++ frontend/app/api/update/meeting/suspend/route.ts
@@ -2,7 +2,7 @@ import { Prisma, Role } from '@prisma/client';
 import { NextResponse, after } from 'next/server';
 import { requireRole } from '../../../../../services/auth';
 import { trimCalendarEventSeries, deleteCalendarEvent, calendarIdsForMeeting } from '../../../../../services/googleCalendar';
-import { formatETDateString, getETDayBounds } from '../../../../../util/timeUtils';
+import { formatETDateString, getETDayBounds } from '../../../../../util/date/timeUtils';
 import { addOneETDay } from '../../../../../util/meetingOccurrences';
 import { getUnresolvedSuspension, reconcilePendingResume, createPendingResumeSeries, MeetingWithPattern } from '../../../../../util/suspension';
 import { prisma } from '../../../../../lib/prisma';
