+++ frontend/app/components/calendar/desktop/WeekView.tsx
@@ -3,9 +3,9 @@ import styles from '../../../../styles/components/calendar/desktop/WeekView.modu
 import DayColumn from "../shared/DayColumn";
 import { filterMeetingsForDate, MeetingFilters } from "../../../../util/meetingFilters";
 import { ROOM_COLORS, ZOOM_ROOM_COLOR, REMOTE_COLOR } from "../../../../util/filterColors";
-import { formatETDateString } from "../../../../util/timeUtils";
+import { formatETDateString } from "../../../../util/date/timeUtils";
 import { layoutOverlappingMeetings } from "../../../../util/meetingOverlapLayout";
-import { getFirstDayOfWeek, getDaysOfWeek } from "../../../../util/weekDates";
+import { getFirstDayOfWeek, getDaysOfWeek } from "../../../../util/date/weekDates";
 import { useWeekMeetings, WeekMeeting } from "../../../../hooks/useWeekMeetings";
 import TopLoadingBar from "../../atoms/TopLoadingBar";
