+++ frontend/app/components/calendar/mobile/DayPortraitView.tsx
@@ -5,9 +5,9 @@ import CalendarHeader from "../shared/CalendarHeader";
 import DayColumn from "../shared/DayColumn";
 import { filterMeetingsForDate, MeetingFilters } from "../../../../util/meetingFilters";
 import { ROOM_COLORS, ZOOM_ROOM_COLOR, REMOTE_COLOR } from "../../../../util/filterColors";
-import { formatETDateString } from "../../../../util/timeUtils";
+import { formatETDateString } from "../../../../util/date/timeUtils";
 import { layoutOverlappingMeetings, OverlapMeeting } from "../../../../util/meetingOverlapLayout";
-import { getFirstDayOfWeek, addDaysToDate } from "../../../../util/weekDates";
+import { getFirstDayOfWeek, addDaysToDate } from "../../../../util/date/weekDates";
 import { useWeekMeetings } from "../../../../hooks/useWeekMeetings";
 import { useScrollNavHide } from "../../../../hooks/useScrollNavHide";
 import { useCalendarContext } from "../../../context/CalendarProvider";
