+++ frontend/app/components/meeting-form/ResumeMeetingModal.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import styles from '../../../styles/components/meeting-form/ResumeMeetingModal.module.scss';
 import DatePicker from '../atoms/DatePicker';
-import { formatETDateString, parseMMDDYYYY } from '../../../util/timeUtils';
+import { formatETDateString, parseMMDDYYYY } from '../../../util/date/timeUtils';
 
 interface ResumeMeetingModalProps {
   isOpen: boolean;
