+++ frontend/app/context/CalendarProvider.tsx
@@ -2,7 +2,7 @@
 
 import React, { createContext, useCallback, useContext, useEffect, useMemo, useRef, useState } from "react";
 import { createDefaultFilters, MeetingFilters } from "../../util/meetingFilters";
-import { getSwipeDirection, isSameWeek, SwipeDirection } from "../../util/weekStripTransition";
+import { getSwipeDirection, isSameWeek, SwipeDirection } from "../../util/date/weekStripTransition";
 
 // Bridges calendar state between HomePage's page content and the globally-mounted AppNavbar
 // (ClientLayout.tsx renders them as siblings, not parent/child, so AppNavbar can't just read
