+++ frontend/util/date/weekStripTransition.ts
