+++ frontend/util/leaseYearCycles.ts
@@ -1,4 +1,4 @@
-import { formatETDateString } from "./timeUtils";
+import { formatETDateString } from "./date/timeUtils";
 
 // Lease years run Jul 1 -> Jun 30 (matches util/leaseDefaults.ts's default settings), stored as
 // UTC dates like the rest of LeaseSettings.
