+++ frontend/util/meetingOccurrences.ts
@@ -1,4 +1,4 @@
-import { getETDayBounds, convertETToUTC, addDaysToETDateString } from "./timeUtils";
+import { getETDayBounds, convertETToUTC, addDaysToETDateString } from "./date/timeUtils";
 import { prisma } from "../lib/prisma";
 import { PublicMeeting, toPublicMeeting } from "./publicMeeting";
