+++ frontend/util/resourceOverlap.ts
@@ -1,6 +1,6 @@
 import { Prisma } from "@prisma/client";
 import { prisma } from "../lib/prisma";
-import { formatETDateString } from "./timeUtils";
+import { formatETDateString } from "./date/timeUtils";
 import { matchesRecurrencePattern, adjustOccurrenceToDate, isDateSuspended } from "./meetingOccurrences";
 
 type SuspensionWindow = { from: Date; to: Date | null };
