+++ frontend/util/suspension.ts
@@ -1,6 +1,6 @@
 import "server-only";
 import { Meeting, RecurrencePattern, SuspensionPeriod } from "@prisma/client";
-import { formatETDateString } from "./timeUtils";
+import { formatETDateString } from "./date/timeUtils";
 import { isDateSuspended, adjustOccurrenceToDate, firstOccurrenceOnOrAfter } from "./meetingOccurrences";
 import { calendarIdsForMeeting, createCalendarEvent, deleteCalendarEvent } from "../services/googleCalendar";
 import { IMeeting } from "../types/models";
